+++ data-raw/ASGS_Ed3_Non_ABS.R
@@ -1,12 +1,13 @@
-abs_geopackage <- "data-raw/ASGS_Ed3_Non_ABS_Structures_GDA2020_GPKG_updated_2022/ASGS_Ed3_Non_ABS_Structures_GDA2020_updated_2022.gpkg"
+abs_geopackage <- "data-raw/ASGS_Ed3_Non_ABS_Structures_GDA2020_GPKG_updated_2023/ASGS_Ed3_Non_ABS_Structures_GDA2020_updated_2023.gpkg"
 
 if (!file.exists(abs_geopackage)) {
   # https://www.abs.gov.au/statistics/standards/australian-statistical-geography-standard-asgs-edition-3
   message("Downloading the source dataset from the ABS website...")
-  asgs_3e_url <- "https://www.abs.gov.au/statistics/standards/australian-statistical-geography-standard-asgs-edition-3/jul2021-jun2026/access-and-downloads/digital-boundary-files/ASGS_Ed3_Non_ABS_Structures_GDA2020_GPKG_updated_2022.zip"
-  download.file(asgs_3e_url, destfile = "data-raw/ASGS_Ed3_Non_ABS_Structures_GDA2020_GPKG_updated_2022.zip")
+  asgs_3e_url <- "https://www.abs.gov.au/statistics/standards/australian-statistical-geography-standard-asgs-edition-3/jul2021-jun2026/access-and-downloads/digital-boundary-files/ASGS_Ed3_Non_ABS_Structures_GDA2020_updated_2023.zip"
+  options(timeout = max(1000L, getOption("timeout")))
+  download.file(asgs_3e_url, destfile = "data-raw/ASGS_Ed3_Non_ABS_Structures_GDA2020_GPKG_updated_2023.zip")
 
-  unzip("data-raw/ASGS_Ed3_Non_ABS_Structures_GDA2020_GPKG_updated_2022.zip", exdir = "data-raw/ASGS_Ed3_Non_ABS_Structures_GDA2020_GPKG_updated_2022")
+  unzip("data-raw/ASGS_Ed3_Non_ABS_Structures_GDA2020_GPKG_updated_2023.zip", exdir = "data-raw/ASGS_Ed3_Non_ABS_Structures_GDA2020_GPKG_updated_2023")
   stopifnot(file.exists(abs_geopackage))
 }
 
@@ -22,7 +23,7 @@ tolerance_m <- 750L
 # list layers available
 st_layers(abs_geopackage)
 
-lga <- read_sf(abs_geopackage, layer = "LGA_2021_AUST_GDA2020")
+lga <- read_sf(abs_geopackage, layer = "LGA_2023_AUST_GDA2020")
 
 crs_nsw <- sf::st_crs(7844) # GDA2020
 crs_working <- sf::st_crs("+proj=eqc +lat_ts=34 units=m")
@@ -36,7 +37,7 @@ crs_working <- sf::st_crs("+proj=eqc +lat_ts=34 units=m")
 
 # The ABS LGA region of "Unincorporated NSW" includes both Lord Howe Island
 # and the Unincorporated Far West Region. Separate these.
-nsw_hires_ui <- filter(lga, LGA_NAME_2021 == "Unincorporated NSW")
+nsw_hires_ui <- filter(lga, LGA_NAME_2023 == "Unincorporated NSW")
 bb <- st_bbox(nsw_hires_ui)
 bb["xmax"] <- 150
 ufwr_hires <- st_intersection(
@@ -55,7 +56,7 @@ usethis::use_data(lhi, overwrite = TRUE)
 
 # The ABS LGA region of "Unincorp. Other Territories" includes both Norfolk
 # Island and the Jervis Bay Territory, of which we only need the latter.
-aus_hires_ui <- filter(lga, LGA_NAME_2021 == "Unincorp. Other Territories")
+aus_hires_ui <- filter(lga, LGA_NAME_2023 == "Unincorp. Other Territories")
 bb <- st_bbox(aus_hires_ui)
 bb["xmax"] <- 153
 jbt_hires <- st_intersection(
@@ -67,7 +68,7 @@ object.size(jbt)
 usethis::use_data(jbt, overwrite = TRUE)
 
 lga_nsw <- lga |>
-  filter(STATE_NAME_2021 == "New South Wales", LGA_NAME_2021 != "Unincorporated NSW") |>
+  filter(STATE_NAME_2021 == "New South Wales", LGA_NAME_2023 != "Unincorporated NSW") |>
   rbind(ufwr_hires) |>
   st_transform(crs_working) |>
   st_simplify(dTolerance = tolerance_m) |>
@@ -82,7 +83,7 @@ object.size(act)
 usethis::use_data(act, overwrite = TRUE)
 
 nsw_hires <- lga |>
-  filter(STATE_NAME_2021 == "New South Wales", LGA_NAME_2021 != "Unincorporated NSW") |>
+  filter(STATE_NAME_2021 == "New South Wales", LGA_NAME_2023 != "Unincorporated NSW") |>
   rbind(ufwr_hires, jbt_hires) |>
   st_transform(crs_working) |>
   st_union() |>