+++ NEWS.md
@@ -1,3 +1,5 @@
+# nswgeo (development version)
+
 # nswgeo 0.3.3
 
 * Initial CRAN release.