+++ NEWS.md
@@ -5,6 +5,7 @@
 * Added geometries for Lord Howe Island, Australian Capital Territory, and
   Jervis Bay Territory, and a utility function `outline()` for using them.
 * Slightly increase the resolution to 750 m instead of 1 km.
+* Update to the 2023 ASGS release, including the 2023 LGA shapefile.
 
 # nswgeo 0.3.3